+++ CHANGELOG.md
@@ -57,6 +57,7 @@
   feedstock data with the resource data methodology
   [PR 801](https://github.com/NatLabRockies/H2Integrate/pull/801).
 - Added capability to specify demand technology for system-level control, and renamed the framework-derived system-level control classification dict from `slc_config` to `slc_topology` to distinguish it from the user-authored `control_parameters` block. [PR 784](https://github.com/NatLabRockies/H2Integrate/pull/784)
+- Add support for slice notation in technology connections to allow users to connect between variables of different shapes. [PR 774](https://github.com/NatLabRockies/H2Integrate/pull/774)
 
 ## 0.8 [April 15, 2026]